添加SoraAnimationEditor 全新修正更新的lAnm, 改名为SoraSpriteAnimation, 引入Sora, 添加render, 封装文件格式更变, 不再封装贴图数据, 保留链接 修正jsongui的parse, 更变为 * container和window的子控件为 widgets: [ { xxxx }, { xxxx } ] 因为jsoncpp不允许重名属性, 比如两个button之类的, 会有一个无法获取
* widget数组也可作为全局object
* 添加对于listbox和dropdown的解析
* 修正各种解析错误
SoraGUIResponser和listener添加一个全新的反射类型: selection, 对应于listbox和dropdown
xmlgui和jsongui添加对于selection反射的解析
修正SoraSprite setTexture之后颜色会变为白色的bug
稍微更改guichan的listmodel基类, 添加pushElement基函数
修正SoraTextureMap在decRf删除texture后不会删除texture记录的bug
guiwrapper的parseXml和json参数调整为SoraWString
修正SoraD3D9Shader shader版本的问题, 现在强制为ps_2_x和vs_2_x
修正SoraD3D9Shader setTexture无法使用的bug
修正SoraCore releaseTexture不会删除贴图的bug
修正guiimage的getWidth和getHeight为getSpriteWith和getSpriteHeight
SoraSprite添加spriteWidth和spriteHeight单独作为getSpriteWidth和getSpriteHeight的返回值
修正SoraHGERenderer的getTargetTexture返回HTEXTURE的bug
修正Windows下字体搜寻的bug
修正SoraSprite setTexture和init的各种矩阵问题
暂时去掉AudiererSoundFile的unref, 貌似有崩溃问题, 待考证
SoraCanvas去掉render*和setFont函数, 因为就Sora的结构没啥必要, Canvas并非基础可渲染物件, 只是renderTarget的代理
SoraCanvas添加shader和imageeffect相关函数, 和Sprite的一致
SoraFreetypeFont的getHeight调整为基高*1.5, 考虑到p之类的英文字母高度问题
修正一些拼写错误

+++ Plugins/SoraGUIChan/soraguifont.cpp
@@ -76,8 +76,7 @@ namespace gcn
         Color color = graphics->getColor();
 		
 		if(mSoraFont) {
-			mSoraFont->setColor(CARGB(color.a, (float)color.r, (float)color.g, (float)color.b));
-			//mSoraFont->Render(x, y, HGETEXT_LEFT, text.c_str());
+			mSoraFont->setColor(CARGB(color.a, color.r, color.g, color.b));
 			mSoraFont->render((float)x, (float)y, sora::s2ws(text+" \0").c_str());
 		}
     }
